Struct curie::PrefixMapping

pub struct PrefixMapping { /* fields omitted */ }

Maps prefixes to base URIs and allows for the expansion of CURIEs (Compact URIs).

Methods

impl PrefixMapping

pub fn set_default(&mut self, default: &str)

Set a default prefix.

This is used during CURIE expansion when there is no prefix, just a reference value.

pub fn add_prefix(
&mut self,
prefix: &str,
value: &str
) -> Result<(), InvalidPrefixError>

Add a prefix to the mapping.

This allows this prefix to be resolved when a CURIE is expanded.

pub fn remove_prefix(&mut self, prefix: &str)

Remove a prefix from the mapping.

Future calls to expand_curie_string or expand_curie that use this prefix will result in a ExpansionError::Invalid error.

pub fn expand_curie_string(&self, curie: &str) -> Result<String, ExpansionError>

Expand a CURIE, returning a complete IRI.

pub fn expand_curie(&self, curie: &Curie) -> Result<String, ExpansionError>

Expand a parsed Curie, returning a complete IRI.

pub fn mappings(&self) -> Iter<String, String>

Return an iterator over the prefix mappings.

This is useful when testing code that uses this crate.