Struct curie::PrefixMapping

pub struct PrefixMapping<'pm> {
    // some fields omitted
}

Maps prefixes to base URIs and allows for the expansion of CURIEs (Compact URIs).

Methods

impl<'pm> PrefixMapping<'pm>

fn set_default(&mut self, default: &'pm str)

Set a default prefix.

This is used during CURIE expansion when there is no prefix, just a reference value.

fn add_prefix(&mut self, prefix: &'pm str, value: &'pm str)

Add a prefix to the mapping.

This allows this prefix to be resolved when expand is invoked on a CURIE.

fn remove_prefix(&mut self, prefix: &str)

Remove a prefix from the mapping.

Future calls to expand that use this prefix will result in a PrefixMappingError::Invalid error.

fn expand(&self, curie: &str) -> Result<String, PrefixMappingError>

Expand a CURIE, returning a complete IRI.